Move MagneticSpaceGroup interface to dedicated submodule
