+++ docs/api/autodoc/spglib.msg.rst
@@ -1,0 +1,8 @@
+spglib.msg module
+=================
+
+.. automodule:: spglib.msg
+   :members:
+   :private-members:
+   :show-inheritance:
+   :undoc-members:
