+++ python/spglib/__init__.py
@@ -8,22 +8,16 @@
 from ._internal import *  # noqa: F403
 from ._version import __version__, __version_tuple__  # noqa: F401
 from .kpoints import *  # noqa: F403
+from .msg import *  # noqa: F403
 from .reduce import *  # noqa: F403
 
 # fmt: off
 from .spglib import (  # noqa: F401
-    MagneticSpaceGroupType,
     SpaceGroupType,
     SpglibDataset,
-    SpglibMagneticDataset,
     find_primitive,
     get_error_message,
     get_hall_number_from_symmetry,
-    get_magnetic_spacegroup_type,
-    get_magnetic_spacegroup_type_from_symmetry,
-    get_magnetic_symmetry,
-    get_magnetic_symmetry_dataset,
-    get_magnetic_symmetry_from_database,
     get_spacegroup,
     get_spacegroup_type,
     get_spacegroup_type_from_symmetry,
